buffer is fo memory, so uint64_t -> size_t

+++ framework/containers/buffer.c
@@ -18,15 +18,15 @@ void buffer_clear(struct Buffer * buffer) {
 	buffer->count = 0;
 }
 
-void buffer_resize(struct Buffer * buffer, uint64_t target_capacity) {
+void buffer_resize(struct Buffer * buffer, size_t target_capacity) {
 	buffer->capacity = target_capacity;
 	if (buffer->count > target_capacity) { buffer->count = target_capacity; }
 	buffer->data = MEMORY_REALLOCATE_ARRAY(buffer, buffer->data, target_capacity);
 }
 
-static void buffer_ensure_capacity(struct Buffer * buffer, uint64_t target_count) {
+static void buffer_ensure_capacity(struct Buffer * buffer, size_t target_count) {
 	if (buffer->capacity < target_count) {
-		uint64_t const target_capacity = grow_capacity_value_u64(buffer->capacity, target_count - buffer->capacity);
+		size_t const target_capacity = grow_capacity_value_u64(buffer->capacity, target_count - buffer->capacity);
 		buffer_resize(buffer, target_capacity);
 	}
 }
@@ -36,7 +36,7 @@ void buffer_push(struct Buffer * buffer, uint8_t value) {
 	buffer->data[buffer->count++] = value;
 }
 
-void buffer_push_many(struct Buffer * buffer, uint64_t count, uint8_t const * value) {
+void buffer_push_many(struct Buffer * buffer, size_t count, uint8_t const * value) {
 	buffer_ensure_capacity(buffer, buffer->count + count);
 	if (value != NULL) {
 		common_memcpy(
@@ -49,12 +49,12 @@ void buffer_push_many(struct Buffer * buffer, uint64_t count, uint8_t const * va
 }
 
 void buffer_align(struct Buffer * buffer) {
-	uint64_t const target_count = align_u64(buffer->count);
+	size_t const target_count = align_u64(buffer->count);
 	buffer_ensure_capacity(buffer, target_count);
 	buffer->count = target_count;
 }
 
-void buffer_set_many(struct Buffer * buffer, uint64_t index, uint64_t count, uint8_t const * value) {
+void buffer_set_many(struct Buffer * buffer, size_t index, size_t count, uint8_t const * value) {
 	if (index + count > buffer->count) { logger_to_console("out of bounds"); DEBUG_BREAK(); return; }
 	common_memcpy(
 		buffer->data + index,
@@ -68,12 +68,12 @@ uint8_t buffer_pop(struct Buffer * buffer) {
 	return buffer->data[--buffer->count];
 }
 
-uint8_t buffer_peek(struct Buffer const * buffer, uint64_t offset) {
+uint8_t buffer_peek(struct Buffer const * buffer, size_t offset) {
 	if (offset >= buffer->count) { DEBUG_BREAK(); return 0; }
 	return buffer->data[buffer->count - offset - 1];
 }
 
-uint8_t buffer_at(struct Buffer const * buffer, uint64_t index) {
+uint8_t buffer_at(struct Buffer const * buffer, size_t index) {
 	if (index >= buffer->count) { DEBUG_BREAK(); return 0; }
 	return buffer->data[index];
 }
